+++ changelog.md
@@ -1,6 +1,7 @@
 # 0.0.2
 
 * Added configurable overrides
+* Added `considerCanBePlaced` config flag to show relevant outlines for placeable blocks
 
 # 0.0.1